requirements.txt
# Core
fastapi>=0.135.0
uvicorn>=0.42.0
pydantic>=2.12.0
pydantic-settings>=2.13.0
python-dotenv>=1.0.0

# Git
gitpython>=3.1.46

# LLM
openai>=2.28.0
anthropic>=0.85.0
json-repair>=0.39.0

# Code parsing
tree-sitter>=0.25.0
tree-sitter-python>=0.25.0
tree-sitter-javascript>=0.25.0
tree-sitter-typescript>=0.23.2
tree-sitter-java>=0.23.5
tree-sitter-ruby>=0.23.1
tree-sitter-go>=0.25.0
tree-sitter-php>=0.23.0
tree-sitter-rust>=0.23.2
tree-sitter-c>=0.23.4
tree-sitter-cpp>=0.23.4
tree-sitter-c-sharp>=0.23.1
tree-sitter-kotlin>=1.0.0
tree-sitter-scala>=0.23.2
tree-sitter-swift>=0.0.1
tree-sitter-bash>=0.23.3
tree-sitter-haskell>=0.21.0
tree-sitter-elixir>=0.3.4
tree-sitter-lua>=0.2.0
tree-sitter-html>=0.23.2
tree-sitter-css>=0.23.2

# Observability
langsmith>=0.7.0

# MCP client
mcp>=1.26.0

# Concurrency
filelock>=3.25.0

# Testing
pytest>=9.0.0
pytest-asyncio>=1.3.0
httpx>=0.28.0
